requirements.txt
django~=5.1.3
djangorestframework~=3.15.2
djangorestframework-simplejwt
psycopg2
django-cors-headers
django-environ
django-debug-toolbar
channels
asgiref
gunicorn
whitenoise
celery[redis]
pytest
pytest-django
pytest-cov
livereload~=2.6.3
factory-boy
drf-yasg
django-filter
django-redis
graphene-django
django-heroku

attrs~=24.2.0
django-livereload~=1.7